Ahora existe y funciona random.js

+++ change-text.py
@@ -1,7 +1,7 @@
 import os
 
 
-def reemplazar_contenido(plantilla, carpeta_destino):
+def reemplazar_contenido(plantilla, carpeta_destino, archivos_a_excluir):
     # Leer contenido de la plantilla
     with open(plantilla, "r") as file_plantilla:
         contenido_plantilla = file_plantilla.read()
@@ -11,7 +11,7 @@ def reemplazar_contenido(plantilla, carpeta_destino):
         ruta_archivo = os.path.join(carpeta_destino, nombre_archivo)
 
         # Verificar si es un archivo y no un directorio
-        if os.path.isfile(ruta_archivo):
+        if os.path.isfile(ruta_archivo) and nombre_archivo not in archivos_a_excluir:
             # Escribir el contenido de la plantilla en el archivo destino
             with open(ruta_archivo, "w") as file_destino:
                 file_destino.write(contenido_plantilla)
@@ -22,5 +22,11 @@ def reemplazar_contenido(plantilla, carpeta_destino):
 plantilla = "test/test.html"
 carpeta_destino = "pokemon/"
 
+# Lista de archivos a excluir
+archivos_a_excluir = [
+    "pokemon.js",
+    "pokemon.css",
+]  # Agrega los nombres de archivos que no quieres modificar
+
 # Llamar a la función para reemplazar contenido
-reemplazar_contenido(plantilla, carpeta_destino)
+reemplazar_contenido(plantilla, carpeta_destino, archivos_a_excluir)
